+++ CHANGELOG.md
@@ -8,3 +8,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 - data-driven inventory and scenarios
 - Abstracted text output for display in a terminal or in a UI
+- Ability to run a scenario with desired inventory items.